+++ demo/site/src/app/not-found.tsx
@@ -1,0 +1,7 @@
+import * as React from "react";
+
+import { NotFoundContent } from "./not-found-content";
+
+export default function NotFound() {
+    return <NotFoundContent />;
+}
